refactor: 레이어에 해당하는 패키지 이외의 패키지를 util 패키지 하위로 이동
