+++ src/main/java/roomescape/controller/dto/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto.request;
+package roomescape.controller.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
@@ -7,7 +7,7 @@ public record LoginRequest(
         @NotBlank
         @Email
         String email,
-        
+
         @NotBlank
         String password
 ) {
