+++ src/main/java/roomescape/controller/dto/request/ReservationTimeSaveRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto.request;
+package roomescape.controller.dto.request;
 
 import java.time.LocalTime;
 import roomescape.domain.roomescape.ReservationTime;
