+++ src/main/java/roomescape/controller/dto/request/ThemeSaveRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto.request;
+package roomescape.controller.dto.request;
 
 import roomescape.domain.roomescape.Theme;
