+++ src/main/java/roomescape/controller/dto/response/ReservationDeleteResponse.java
@@ -1,4 +1,4 @@
-package roomescape.dto.response;
+package roomescape.controller.dto.response;
 
 public record ReservationDeleteResponse(
         int updateCount
