+++ src/main/java/roomescape/controller/dto/response/ReservationResponse.java
@@ -1,4 +1,4 @@
-package roomescape.dto.response;
+package roomescape.controller.dto.response;
 
 import java.time.LocalDate;
 import roomescape.domain.roomescape.Reservation;
