+++ src/main/java/roomescape/controller/dto/response/SelectableTimeResponse.java
@@ -1,7 +1,6 @@
-package roomescape.dto.response;
+package roomescape.controller.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-
 import java.time.LocalTime;
 
 public record SelectableTimeResponse(
