+++ src/main/java/roomescape/service/roomescape/ThemeService.java
@@ -4,10 +4,10 @@ import java.time.LocalDate;
 import java.util.List;
 import java.util.NoSuchElementException;
 import org.springframework.stereotype.Service;
+import roomescape.controller.dto.request.ThemeSaveRequest;
+import roomescape.controller.dto.response.ThemeDeleteResponse;
+import roomescape.controller.dto.response.ThemeResponse;
 import roomescape.domain.roomescape.Theme;
-import roomescape.dto.request.ThemeSaveRequest;
-import roomescape.dto.response.ThemeDeleteResponse;
-import roomescape.dto.response.ThemeResponse;
 import roomescape.repository.roomescape.ReservationDao;
 import roomescape.repository.roomescape.ThemeDao;
