+++ src/main/java/roomescape/util/configuration/AuthenticationPrincipal.java
@@ -1,4 +1,4 @@
-package roomescape.configuration;
+package roomescape.util.configuration;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
