+++ src/main/java/roomescape/util/configuration/AuthenticationPrincipalArgumentResolver.java
@@ -1,4 +1,4 @@
-package roomescape.configuration;
+package roomescape.util.configuration;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
