+++ src/test/java/roomescape/controller/auth/TokenAuthControllerTest.java
@@ -10,8 +10,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.dto.request.LoginRequest;
-import roomescape.dto.request.SignupRequest;
+import roomescape.controller.dto.request.LoginRequest;
+import roomescape.controller.dto.request.SignupRequest;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
