+++ src/test/java/roomescape/controller/roomescape/ReservationControllerTest.java
@@ -17,8 +17,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.dto.response.ReservationResponse;
-import roomescape.dto.response.SelectableTimeResponse;
+import roomescape.controller.dto.response.ReservationResponse;
+import roomescape.controller.dto.response.SelectableTimeResponse;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
